% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_filter}
\alias{param_filter}
\title{Filtering parameters for Zotero API}
\usage{
param_filter(
  items = NULL,
  item_type = NULL,
  since = NULL,
  include_trashed = NULL
)
}
\arguments{
\item{items}{A character vector of item keys. Valid only for item requests.
You can specify up to 50 item keys in a single request.}

\item{item_type}{A character vector specifying item types. Supports Boolean
searches (AND, OR, NOT). See the \verb{Boolean Searches} section for details.}

\item{since}{An integer representing a specific library version. Only items
modified after the specified version (from a previous
\strong{Last-Modified-Version} header) will be returned.}

\item{include_trashed}{Include items in the trash, only valid for Items
Endpoints.}
}
\description{
Filtering parameters for Zotero API
}
\section{Boolean searches}{

\itemize{
\item \code{item_type = "book"}
\item \code{item_type = "book || journalArticle"} (OR)
\item \code{item_type = "-attachment"} (NOT)
}
}

