% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_format}
\alias{param_format}
\title{Format parameters for Zotero API}
\usage{
param_format(
  format = NULL,
  includes = NULL,
  contents = NULL,
  style = NULL,
  linkwrap = NULL,
  locale = NULL
)
}
\arguments{
\item{format}{Format of the response.}

\item{includes}{Formats to include in the response, multiple formats can be
specified.}

\item{contents}{The format of the Atom response's \if{html}{\out{<content>}} node, multiple
formats can be specified.}

\item{style}{Citation style for formatted references. You can provide either
the name of a style (e.g., \code{"apa"}) or a URL to a custom CSL file. Only valid
when \code{format = "bib"}, or when \code{includes} or \code{contents} contains \code{"bib"} or
\code{"citation"}.}

\item{linkwrap}{A boolean indicating whether URLs and DOIs should be returned
as links. Only valid when \code{format = "bib"}, or when \code{includes} or \code{contents}
contains \code{"bib"} or \code{"citation"}.}

\item{locale}{A character string specifying the locale to use for
bibliographic formatting (e.g., \code{"en-US"}). Only valid when \code{format = "bib"},
or when \code{includes} or \code{contents} contains \code{"bib"} or \code{"citation"}.}
}
\description{
Format parameters for Zotero API
}
