% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_pagination}
\alias{param_pagination}
\title{Pagination and Soring parameters for Zotero API}
\usage{
param_pagination(start = NULL, limit = NULL, sort = NULL, direction = NULL)
}
\arguments{
\item{start}{The index of the first result. Combine with the \code{limit}
parameter to select a slice of the available results.}

\item{limit}{The maximum number of results to return with a single
request. Required for export formats. An integer between 1-100.}

\item{sort}{The name of the field by which entries are sorted.}

\item{direction}{The sorting direction of the field specified in the
sort parameter. One of \code{"asc"} or \code{"desc"}.}
}
\description{
Pagination and Soring parameters for Zotero API
}
