% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{param_search}
\alias{param_search}
\title{Searching parameters for Zotero API}
\usage{
param_search(quick = NULL, mode = NULL, tag = NULL)
}
\arguments{
\item{quick}{A character string for a quick search. Use the \code{mode} parameter
to change the search mode. Currently, only phrase searching is supported.}

\item{mode}{A character string specifying the search mode:
\itemize{
\item For \strong{items} endpoint, you can use one of the following:
\itemize{
\item \code{"titleCreatorYear"}: Search by title, creator, and year.
\item \code{"everything"}: Search across all fields for items.
}
\item For \strong{tags} endpoint, you can use one of the following:
\itemize{
\item \code{"contains"}: Tag search mode where the query string must be contained in
the tag.
\item \code{"startsWith"}: Tag search mode where the query string must match the
beginning of the tag.
}
}}

\item{tag}{A character vector specifying the tags. Supports Boolean searches
(AND, OR, NOT). See the \verb{Boolean Searches} section for details.}
}
\description{
Searching parameters for Zotero API
}
\section{Boolean searches}{

\itemize{
\item \code{tag = "foo"}
\item \code{tag = "foo bar"} (tag with space)
\item \code{tag = c("foo", "bar")}: Equivalent to\code{"tag=foo&tag=bar"} (AND)
\item \code{tag = "foo bar || bar"} (OR)
\item \code{tag = "-foo"} (NOT)
\item \code{tag = "\\-foo"} (literal first-character hyphen)
}
}

