% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{zotero_autho}
\alias{zotero_autho}
\title{Zotero API Key Management and OAuth Authorization}
\usage{
zotero_autho(api_key = NULL, oauth_userid = NULL)
}
\arguments{
\item{api_key}{Optional API key. If provided, the function will set the API
key in the memory and return without performing OAuth authorization.}

\item{oauth_userid}{Optional user ID. Used to retrieve a cached OAuth token
specific to the user. If provided, the function will search for a cached
token associated with this user ID. If not provided, the function will use
the last used token key file or the most recently available cached token.}
}
\description{
This function manages API key handling and OAuth authorization for the Zotero
API.
}
