% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{zotero_key}
\alias{zotero_key}
\title{Retrieve the global Zotero API Key}
\usage{
zotero_key()
}
\description{
This function returns the currently cached Zotero API key from the internal
cache. The key is used for authenticated requests to the Zotero API.
}
