% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zotero_library}
\alias{zotero_library}
\title{Zotero Library}
\usage{
zotero_library(library_id, library_type)
}
\arguments{
\item{library_id}{A string representing either a \verb{user ID} or \verb{group ID} for
the library. This is the unique identifier for the Zotero library.}

\item{library_type}{A string that specifies the type of the library. It can
either be:
\itemize{
\item \code{"user"}: Represents a user library.
\item \code{"group"}: Represents a group library.
}}
}
\value{
A \code{zotero_library} object.
}
\description{
This function creates a \code{zotero_library} object, which represents a specific
Zotero library.  A library is identified by a \verb{user ID} or \verb{group ID}, and
its type is either \code{"user"} or \code{"group"}. The \code{zotero_library} object is used
when specifying the library for API requests to \link{Zotero} instance, ensuring
that the request is scoped to a specific user or group library.
}
