% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zotero_params}
\alias{zotero_params}
\title{Parameters for Zotero API}
\usage{
zotero_params(
  sort = NULL,
  direction = NULL,
  limit = NULL,
  start = NULL,
  item_search = NULL,
  tag_search = NULL,
  format = NULL,
  format_includes = NULL,
  format_contents = NULL,
  style = NULL,
  linkwrap = NULL,
  locale = NULL
)
}
\arguments{
\item{sort}{The name of the field by which entries are sorted.}

\item{direction}{The sorting direction of the field specified in the
sort parameter. One of \code{"asc"} or \code{"desc"}.}

\item{limit}{The maximum number of results to return with a single
request. Required for export formats. An integer between 1-100.}

\item{start}{The index of the first result. Combine with the \code{limit}
parameter to select a slice of the available results.}

\item{item_search, tag_search}{A \code{\link[=zotero_search]{zotero_search()}} object to
refine the item/tag searching.}

\item{format}{Format of the response.}

\item{format_includes}{Formats to include in the response, multiple
formats can be specified.}

\item{format_contents}{The format of the Atom response's \if{html}{\out{<content>}}
node, multiple formats can be specified.}

\item{style}{Citation style for formatted references. You can provide
either the name of a style (e.g., \code{"apa"}) or a URL to a custom CSL
file. Only valid when \code{format = "bib"}, or when \code{format_includes} or
\code{format_contents} contains \code{"bib"} or \code{"citation"}.}

\item{linkwrap}{A boolean indicating whether URLs and DOIs should be
returned as links. Only valid when \code{format = "bib"}, or when
\code{format_includes} or \code{format_contents} contains \code{"bib"} or
\code{"citation"}.}

\item{locale}{A character string specifying the locale to use for
bibliographic formatting (e.g., \code{"en-US"}). Only valid when \code{format = "bib"}, or when \code{format_includes} or \code{format_contents} contains
\code{"bib"} or \code{"citation"}.}
}
\description{
Parameters for Zotero API
}
