% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zotero_perform}
\alias{zotero_perform}
\title{Perform the Zotero API Request}
\usage{
zotero_perform(req, ..., key = NULL)
}
\arguments{
\item{req}{A httr2 \link[httr2:request]{request} or a list of
\link[httr2:request]{request} objects created by \code{\link{Zotero}}.}

\item{...}{Additional arguments passed to method dispatch.
\itemize{
\item When \code{req} is a single request, these arguments are passed to
\code{\link[httr2:req_perform]{httr2::req_perform()}}.
\item When \code{req} is a list of requests, these arguments are passed to
\code{\link[httr2:req_perform_parallel]{httr2::req_perform_parallel()}}.
}}

\item{key}{Optional authenticated key for the request.}
}
\value{
The response from the Zotero API.
}
\description{
This function executes a Zotero API request created by \code{\link{Zotero}}. It uses
method dispatch to handle different types of request execution based on the
input. If a single request is provided, it performs the request. If a list of
requests is provided, it performs the requests in parallel.
}
