% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zotero_req_groups}
\alias{zotero_req_groups}
\title{Zotero Groups Access Information}
\usage{
zotero_req_groups(userid)
}
\arguments{
\item{userid}{A single string representing the \code{userID} or \code{groupID}.  User
IDs differ from usernames and can be obtained from the \href{https://www.zotero.org/settings/keys}{\verb{API Keys}} page or \code{\link[=zotero_req_key]{zotero_req_key()}}.}
}
\description{
This function retrieves the set of groups the current API key has access to,
including public groups the key owner belongs to, even if the key doesn't
have explicit permissions for those groups. It is useful for identifying
which groups a user or API key can interact with, including public groups and
groups that the owner has membership in.
}
