% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zotero_req_key}
\alias{zotero_req_key}
\title{Zotero API Key Privileges and User Information}
\usage{
zotero_req_key(key = NULL)
}
\arguments{
\item{key}{otero API Key for the request. By default, the function will
attempt to use the \link[=zotero_key]{global API key}.}
}
\description{
This function retrieves the user ID and privileges associated with a given
Zotero API key.
}
