% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zotero_req_search}
\alias{zotero_req_search}
\title{Zotero Search Metadata Request}
\usage{
zotero_req_search(prefix, search = NULL)
}
\arguments{
\item{prefix}{A single string representing either a \code{userID} or \code{groupID}.
User IDs differ from usernames and can be obtained from the
\href{https://www.zotero.org/settings/keys}{\verb{API Keys}} page or
\code{\link[=zotero_req_key]{zotero_req_key()}}. Group IDs are different from group names and can be
retrieved from \code{\link[=zotero_req_groups]{zotero_req_groups()}}.}

\item{search}{A specific saved search in the library. If \code{NULL}, the function
will get all saved searches in the library.}
}
\description{
Zotero Search Metadata Request
}
\note{
Only search metadata is currently available, not search results.
}
