% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zotero_request}
\alias{zotero_request}
\title{Make a Zotero API Request}
\usage{
zotero_request(..., query = NULL)
}
\arguments{
\item{...}{Character strings representing the path components to append
to the base URL (\verb{https://api.zotero.org}).}

\item{query}{Optional named list of query parameters to add to the request
URL.  Defaults to \code{NULL}.}
}
\value{
A \code{httr2} request object that can be passed to \code{zotero_perform} for
execution.
}
\description{
This function constructs a request to the Zotero API. It allows appending
paths to the base URL and optionally adding query parameters.
}
\seealso{
\url{https://www.zotero.org/support/dev/web_api/v3/basics}
}
