% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zotero_search}
\alias{zotero_search}
\title{Searching Parameters for Zotero API}
\usage{
zotero_search(
  quick = NULL,
  mode = NULL,
  include_tags = NULL,
  include_items = NULL,
  include_item_types = NULL,
  since = NULL
)
}
\arguments{
\item{quick}{A character string for a quick search. Use the \code{mode} parameter
to change the search mode. Currently, only phrase searching is supported.}

\item{mode}{A character string specifying the search mode:
\itemize{
\item For \strong{items} endpoint, you can use one of the following:
\itemize{
\item \code{"titleCreatorYear"}: Search by title, creator, and year.
\item \code{"everything"}: Search across all fields for items.
}
\item For \strong{tags} endpoint, you can use one of the following:
\itemize{
\item \code{"contains"}: Tag search mode where the query string must be contained in
the tag.
\item \code{"startsWith"}: Tag search mode where the query string must match the
beginning of the tag.
}
}}

\item{include_tags}{A character vector specifying the tags. Supports Boolean
searches (AND, OR, NOT). See the \verb{Boolean Searches} section for details.}

\item{include_items}{A character vector of item keys. Valid only for item
requests. You can specify up to 50 item keys in a single request.}

\item{include_item_types}{A character vector specifying item types. Supports
Boolean searches (AND, OR, NOT). See the \verb{Boolean Searches} section for
details.}

\item{since}{An integer representing a specific library version. Only items
modified after the specified version (from a previous
\strong{Last-Modified-Version} header) will be returned.}
}
\description{
Searching Parameters for Zotero API
}
\section{Boolean searches}{

\itemize{
\item \code{include_item_types = "book"}
\item \code{include_item_types = "book || journalArticle"} (OR)
\item \code{include_item_types = "-attachment"} (NOT)
\item \code{include_tags = "foo"}
\item \code{include_tags = "foo bar"} (tag with space)
\item \code{include_tags = c("foo", "bar")}: Equivalent to\code{"tag=foo&tag=bar"} (AND)
\item \code{include_tags = "foo bar || bar"} (OR)
\item \code{include_tags = "-foo"} (NOT)
\item \code{include_tags = "\\-foo"} (literal first-character hyphen)
}
}

