% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero-package.R
\docType{package}
\name{zotero-package}
\alias{zotero}
\alias{zotero-package}
\title{zotero: An API Client for the Zotero}
\description{
R wrapper for the Zotero API.
}
\author{
\strong{Maintainer}: Yun Peng \email{pengyun.me@outlook.com} (\href{https://orcid.org/0000-0003-2801-3332}{ORCID})

}
\keyword{internal}
