% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Zotero}
\alias{Zotero}
\title{Zotero instance}
\description{
Zotero instance

Zotero instance
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Zotero-key_set}{\code{Zotero$key_set()}}
\item \href{#method-Zotero-key_oauth}{\code{Zotero$key_oauth()}}
\item \href{#method-Zotero-key_read}{\code{Zotero$key_read()}}
\item \href{#method-Zotero-key_cache}{\code{Zotero$key_cache()}}
\item \href{#method-Zotero-key_complete_info}{\code{Zotero$key_complete_info()}}
\item \href{#method-Zotero-key_userid}{\code{Zotero$key_userid()}}
\item \href{#method-Zotero-key_username}{\code{Zotero$key_username()}}
\item \href{#method-Zotero-key_access}{\code{Zotero$key_access()}}
\item \href{#method-Zotero-key_groups}{\code{Zotero$key_groups()}}
\item \href{#method-Zotero-key_revoke}{\code{Zotero$key_revoke()}}
\item \href{#method-Zotero-library}{\code{Zotero$library()}}
\item \href{#method-Zotero-request}{\code{Zotero$request()}}
\item \href{#method-Zotero-searches}{\code{Zotero$searches()}}
\item \href{#method-Zotero-clone}{\code{Zotero$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_set"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_set}{}}}
\subsection{Method \code{key_set()}}{
Set Zotero API key directly
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_set(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{A string representing the Zotero API key. If provided.
Visit \url{https://www.zotero.org/settings/keys} to create a private
API key.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_oauth"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_oauth}{}}}
\subsection{Method \code{key_oauth()}}{
OAuth authorization for the Zotero API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_oauth()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method must be used in an interactive session
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_read"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_read}{}}}
\subsection{Method \code{key_read()}}{
Read the cached Zotero API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_read(userid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{A string of user ID. If provided, the function will
attempt to find the cached key for this user. If not provided, the
function will attempt to use the last key used. For your user ID,
visit \url{https://www.zotero.org/settings/keys}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_cache"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_cache}{}}}
\subsection{Method \code{key_cache()}}{
Cache the Zotero API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_complete_info"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_complete_info}{}}}
\subsection{Method \code{key_complete_info()}}{
Fill in missing information associated with the Zotero
API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_complete_info()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method ensures that the Zotero API key is fully populated with
the required user information (e.g., userid, username, and access
level). If any of the key details are missing, it retrieves and
completes the necessary information.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_userid"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_userid}{}}}
\subsection{Method \code{key_userid()}}{
Retrieve the Zotero user ID for the current API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_userid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_username"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_username}{}}}
\subsection{Method \code{key_username()}}{
Retrieve the Zotero user name for the current API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_username()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_access"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_access}{}}}
\subsection{Method \code{key_access()}}{
Retrieve the access level for the current API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_access()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_groups"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_groups}{}}}
\subsection{Method \code{key_groups()}}{
Retrieve the list of groups the current API key has
access to
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_groups()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method retrieves the set of groups the current API key has
access to, including public groups the key owner belongs to, even if
the key doesn't have explicit permissions for those groups. It is
useful for identifying which groups a user or API key can interact
with, including public groups and groups that the owner has
membership in.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_revoke"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_revoke}{}}}
\subsection{Method \code{key_revoke()}}{
Revoke the current Zotero API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_revoke()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method will also remove the cached file if the key is from the
cached credential file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-library"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-library}{}}}
\subsection{Method \code{library()}}{
Determine or Set the Library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$library(library_id, library_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{library_id}}{A string representing either a \verb{user ID} or
\verb{group ID} for the library.}

\item{\code{library_type}}{A string of \code{"user"} or \code{"group"} representing
the type for the library.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method allows you to specify the library from which to retrieve
or manipulate data. The library ID is identified by a \verb{user ID} or
\verb{group ID}. The library ID and its type (\code{user} or \code{group}) are
essential for the request. By default, the \verb{user ID} associated with
the current API key is used.

User IDs differ from usernames and can be found on the \href{https://www.zotero.org/settings/keys}{API Keys} page or retrieved using
the \verb{$key_userid()} method. Group IDs are distinct from group names
and can be obtained via the \verb{$key_groups()} method.

If no \code{library_id} is provided, the method will return the currently
used library. If \code{library_id} is set to \code{NULL}, the library will be
reset to the default.  Otherwise, if a \code{library_id} is provided,
\code{library_type} will default to \code{"group"}, as it is uncommon for users
to set a \code{user} type with a library ID that differs from the current
API key's user ID.  If a \code{library_id} is provided with a non-default
\code{library_type}, the \code{library_type} (\code{"user"} or \code{"group"}) must be
specified.
}

\subsection{Returns}{
A \code{zotero_library} object when \code{library_id} is missing,
otherwise, returns the Zotero instance itself, allowing for method
chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-request"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-request}{}}}
\subsection{Method \code{request()}}{
Make a request to the Zotero API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$request(..., query = NULL, prefix = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Character strings representing the path components to
append to the Zotero API base URL.}

\item{\code{query}}{Optional named list of query parameters to be added to
the request URL.}

\item{\code{prefix}}{Whether to include the library-specific prefix. If
\code{TRUE}, the request URL will be prefixed with the relevant library
path (\verb{/users/<userID>} or \verb{/groups/<groupID>}), ensuring that the
request is scoped to a specific library.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method constructs a request of the Zotero API. It allows you to
specify the URL path components and optional query parameters, and
includes support for prefixing library-specific paths (e.g.,
\verb{/users/<userID>} or \verb{/groups/<groupID>}) when interacting with data
from specific libraries.
}

\subsection{Returns}{
A \code{httr2} \link[httr2:request]{request} object that can be used
to send the API request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-searches"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-searches}{}}}
\subsection{Method \code{searches()}}{
Get saved searches from Zotero API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$searches(search_key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search_key}}{A specific saved search in the library. If \code{NULL},
the method will get all saved searches in the library.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Only get the saved searches, not search results.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
