% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{Zotero}
\alias{Zotero}
\title{Zotero instance}
\description{
Zotero instance

Zotero instance
}
\seealso{
A full list of methods can be found here:
\url{http://www.zotero.org/support/dev/server_api}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Zotero-key_set}{\code{Zotero$key_set()}}
\item \href{#method-Zotero-key_oauth}{\code{Zotero$key_oauth()}}
\item \href{#method-Zotero-key_read}{\code{Zotero$key_read()}}
\item \href{#method-Zotero-key_cache}{\code{Zotero$key_cache()}}
\item \href{#method-Zotero-key_complete_info}{\code{Zotero$key_complete_info()}}
\item \href{#method-Zotero-key_userid}{\code{Zotero$key_userid()}}
\item \href{#method-Zotero-key_username}{\code{Zotero$key_username()}}
\item \href{#method-Zotero-key_access}{\code{Zotero$key_access()}}
\item \href{#method-Zotero-key_groups}{\code{Zotero$key_groups()}}
\item \href{#method-Zotero-key_revoke}{\code{Zotero$key_revoke()}}
\item \href{#method-Zotero-library}{\code{Zotero$library()}}
\item \href{#method-Zotero-parameters}{\code{Zotero$parameters()}}
\item \href{#method-Zotero-request}{\code{Zotero$request()}}
\item \href{#method-Zotero-collections}{\code{Zotero$collections()}}
\item \href{#method-Zotero-collections_top}{\code{Zotero$collections_top()}}
\item \href{#method-Zotero-collection}{\code{Zotero$collection()}}
\item \href{#method-Zotero-collection_subgroups}{\code{Zotero$collection_subgroups()}}
\item \href{#method-Zotero-collection_tags}{\code{Zotero$collection_tags()}}
\item \href{#method-Zotero-collection_items}{\code{Zotero$collection_items()}}
\item \href{#method-Zotero-collection_items_tags}{\code{Zotero$collection_items_tags()}}
\item \href{#method-Zotero-collection_items_top}{\code{Zotero$collection_items_top()}}
\item \href{#method-Zotero-collection_items_top_tags}{\code{Zotero$collection_items_top_tags()}}
\item \href{#method-Zotero-items}{\code{Zotero$items()}}
\item \href{#method-Zotero-items_tags}{\code{Zotero$items_tags()}}
\item \href{#method-Zotero-items_top}{\code{Zotero$items_top()}}
\item \href{#method-Zotero-items_top_tags}{\code{Zotero$items_top_tags()}}
\item \href{#method-Zotero-items_trash}{\code{Zotero$items_trash()}}
\item \href{#method-Zotero-items_trash_tags}{\code{Zotero$items_trash_tags()}}
\item \href{#method-Zotero-item}{\code{Zotero$item()}}
\item \href{#method-Zotero-item_tags}{\code{Zotero$item_tags()}}
\item \href{#method-Zotero-item_children}{\code{Zotero$item_children()}}
\item \href{#method-Zotero-publication_items}{\code{Zotero$publication_items()}}
\item \href{#method-Zotero-publication_items_tags}{\code{Zotero$publication_items_tags()}}
\item \href{#method-Zotero-saved_searches}{\code{Zotero$saved_searches()}}
\item \href{#method-Zotero-saved_search}{\code{Zotero$saved_search()}}
\item \href{#method-Zotero-tags}{\code{Zotero$tags()}}
\item \href{#method-Zotero-clone}{\code{Zotero$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_set"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_set}{}}}
\subsection{Method \code{key_set()}}{
Set Zotero API key directly
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_set(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{A character string representing the Zotero API key. You
can create your private API key by visiting
\url{https://www.zotero.org/settings/keys}. If the \code{key} argument is
not provided, the function will attempt to use the value from the
environment variable \code{ZOTERO_API}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_oauth"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_oauth}{}}}
\subsection{Method \code{key_oauth()}}{
OAuth authorization for the Zotero API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_oauth()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method must be used in an interactive session
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_read"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_read}{}}}
\subsection{Method \code{key_read()}}{
Read the cached Zotero API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_read(userid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userid}}{A string of user ID. If provided, the function will
attempt to find the cached key for this user. If not provided, the
function will attempt to use the last key used. For your user ID,
visit \url{https://www.zotero.org/settings/keys}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_cache"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_cache}{}}}
\subsection{Method \code{key_cache()}}{
Cache the Zotero API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_cache()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_complete_info"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_complete_info}{}}}
\subsection{Method \code{key_complete_info()}}{
Fill in missing information associated with the Zotero
API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_complete_info()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method ensures that the Zotero API key is fully populated with
the required user information (e.g., userid, username, and access
level). If any of the key details are missing, it retrieves and
completes the necessary information.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_userid"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_userid}{}}}
\subsection{Method \code{key_userid()}}{
Retrieve the Zotero user ID for the current API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_userid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_username"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_username}{}}}
\subsection{Method \code{key_username()}}{
Retrieve the Zotero user name for the current API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_username()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_access"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_access}{}}}
\subsection{Method \code{key_access()}}{
Retrieve the access level for the current API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_access()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_groups"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_groups}{}}}
\subsection{Method \code{key_groups()}}{
Retrieve the list of groups the current API key has
access to
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_groups()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method retrieves the set of groups the current API key has
access to, including public groups the key owner belongs to, even if
the key doesn't have explicit permissions for those groups. It is
useful for identifying which groups a user or API key can interact
with, including public groups and groups that the owner has
membership in.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-key_revoke"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-key_revoke}{}}}
\subsection{Method \code{key_revoke()}}{
Revoke the current Zotero API key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$key_revoke()}\if{html}{\out{</div>}}
}

\subsection{Details}{
This method will also remove the cached file if the key is from the
cached credential file.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-library"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-library}{}}}
\subsection{Method \code{library()}}{
Determine or Set the Library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$library(library_id, library_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{library_id}}{A string representing either a \verb{user ID} or
\verb{group ID} for the library.}

\item{\code{library_type}}{A string of \code{"user"} or \code{"group"} representing
the type for the library.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method allows you to specify the library from which to retrieve
or manipulate data. The library ID is identified by a \verb{user ID} or
\verb{group ID}. The library ID and its type (\code{user} or \code{group}) are
essential for the request. By default, the \verb{user ID} associated with
the current API key is used.

User IDs differ from usernames and can be found on the \href{https://www.zotero.org/settings/keys}{API Keys} page or retrieved using
the \verb{$key_userid()} method. Group IDs are distinct from group names
and can be obtained via the \verb{$key_groups()} method.

If no \strong{library_id} is provided, the method will return the
currently used library. If \strong{library_id} is set to \code{NULL}, the
library will be reset to the default.  Otherwise, if a \strong{library_id}
is provided, \strong{library_type} will default to \code{"group"}, as it is
uncommon for users to set a \code{user} type with a library ID that
differs from the current API key's user ID.  If a \strong{library_id} is
provided with a non-default \strong{library_type}, the \strong{library_type}
(\code{"user"} or \code{"group"}) must be specified.
}

\subsection{Returns}{
A \code{zotero_library} object when \strong{library_id} is missing,
otherwise, returns the Zotero instance itself, allowing for method
chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-parameters"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-parameters}{}}}
\subsection{Method \code{parameters()}}{
Determine or Set the global searching parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$parameters(
  sort = NULL,
  direction = NULL,
  limit = NULL,
  start = NULL,
  item_search = NULL,
  tag_search = NULL,
  format = NULL,
  format_includes = NULL,
  format_contents = NULL,
  style = NULL,
  linkwrap = NULL,
  locale = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sort}}{The name of the field by which entries are sorted.}

\item{\code{direction}}{The sorting direction of the field specified in the
sort parameter. One of \code{"asc"} or \code{"desc"}.}

\item{\code{limit}}{The maximum number of results to return with a single
request. Required for export formats. An integer between 1-100.}

\item{\code{start}}{The index of the first result. Combine with the \code{limit}
parameter to select a slice of the available results.}

\item{\code{item_search, tag_search}}{A \code{\link[=zotero_search]{zotero_search()}} object to
refine the item/tag searching.}

\item{\code{format}}{Format of the response.}

\item{\code{format_includes}}{Formats to include in the response, multiple
formats can be specified.}

\item{\code{format_contents}}{The format of the Atom response's \if{html}{\out{<content>}}
node, multiple formats can be specified.}

\item{\code{style}}{Citation style for formatted references. You can provide
either the name of a style (e.g., \code{"apa"}) or a URL to a custom CSL
file. Only valid when \code{format = "bib"}, or when \code{format_includes} or
\code{format_contents} contains \code{"bib"} or \code{"citation"}.}

\item{\code{linkwrap}}{A boolean indicating whether URLs and DOIs should be
returned as links. Only valid when \code{format = "bib"}, or when
\code{format_includes} or \code{format_contents} contains \code{"bib"} or
\code{"citation"}.}

\item{\code{locale}}{A character string specifying the locale to use for
bibliographic formatting (e.g., \code{"en-US"}). Only valid when \code{format = "bib"}, or when \code{format_includes} or \code{format_contents} contains
\code{"bib"} or \code{"citation"}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-request"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-request}{}}}
\subsection{Method \code{request()}}{
Make a request to the Zotero API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$request(..., library = NULL, query = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Character strings representing the path components to
append to the Zotero API base URL.}

\item{\code{library}}{A \code{\link[=zotero_library]{zotero_library()}} object. If provided, the
request URL will be prefixed with the relevant library path
(\verb{/users/<userID>} or \verb{/groups/<groupID>}), ensuring that the request
is scoped to a specific library.}

\item{\code{query}}{Optional named list of query parameters to be added to
the request URL.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method constructs a request of the Zotero API. It allows you to
specify the URL path components and optional query parameters, and
includes support for prefixing library-specific paths (e.g.,
\verb{/users/<userID>} or \verb{/groups/<groupID>}) when interacting with data
from specific libraries.
}

\subsection{Returns}{
A \code{httr2} \link[httr2:request]{request} object that can be used
to send the API request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-collections"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-collections}{}}}
\subsection{Method \code{collections()}}{
Collections in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$collections()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-collections_top"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-collections_top}{}}}
\subsection{Method \code{collections_top()}}{
Top-level collections in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$collections_top()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-collection"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-collection}{}}}
\subsection{Method \code{collection()}}{
A specific collection in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$collection(collection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{A specific collection in the library.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-collection_subgroups"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-collection_subgroups}{}}}
\subsection{Method \code{collection_subgroups()}}{
Subcollections within a specific collection in the
library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$collection_subgroups(collection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{A specific collection in the library.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-collection_tags"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-collection_tags}{}}}
\subsection{Method \code{collection_tags()}}{
Tags associated with a specific collection in the
library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$collection_tags(collection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{A specific collection in the library.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-collection_items"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-collection_items}{}}}
\subsection{Method \code{collection_items()}}{
Items within a specific collection in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$collection_items(collection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{A specific collection in the library.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-collection_items_tags"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-collection_items_tags}{}}}
\subsection{Method \code{collection_items_tags()}}{
Tags associated with the Items within a specific
collection in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$collection_items_tags(collection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{A specific collection in the library.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-collection_items_top"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-collection_items_top}{}}}
\subsection{Method \code{collection_items_top()}}{
Top-level items within a specific collection in the
library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$collection_items_top(collection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{A specific collection in the library.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-collection_items_top_tags"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-collection_items_top_tags}{}}}
\subsection{Method \code{collection_items_top_tags()}}{
Tags associated with the top-level items within a
specific collection in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$collection_items_top_tags(collection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collection}}{A specific collection in the library.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-items"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-items}{}}}
\subsection{Method \code{items()}}{
All items in the library, excluding trashed items
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$items()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-items_tags"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-items_tags}{}}}
\subsection{Method \code{items_tags()}}{
Tags associated all items in the library, excluding
trashed items
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$items_tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-items_top"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-items_top}{}}}
\subsection{Method \code{items_top()}}{
Top-level items in the library, excluding trashed items
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$items_top()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-items_top_tags"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-items_top_tags}{}}}
\subsection{Method \code{items_top_tags()}}{
Tags associated with the top-level items in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$items_top_tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-items_trash"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-items_trash}{}}}
\subsection{Method \code{items_trash()}}{
Items in the trash
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$items_trash()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-items_trash_tags"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-items_trash_tags}{}}}
\subsection{Method \code{items_trash_tags()}}{
Tags associated with the items in the trash
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$items_trash_tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-item"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-item}{}}}
\subsection{Method \code{item()}}{
A specific item in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$item(item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{A specific saved search in the library.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-item_tags"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-item_tags}{}}}
\subsection{Method \code{item_tags()}}{
Tags associated with a specific item in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$item_tags(item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{A specific saved search in the library.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-item_children"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-item_children}{}}}
\subsection{Method \code{item_children()}}{
Child items under a specific item
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$item_children(item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{A specific saved search in the library.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-publication_items"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-publication_items}{}}}
\subsection{Method \code{publication_items()}}{
Items in My Publications
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$publication_items()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-publication_items_tags"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-publication_items_tags}{}}}
\subsection{Method \code{publication_items_tags()}}{
Tags associated with the items in My Publications
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$publication_items_tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-saved_searches"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-saved_searches}{}}}
\subsection{Method \code{saved_searches()}}{
Get all saved searches in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$saved_searches()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Only get the saved searches, not search results.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-saved_search"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-saved_search}{}}}
\subsection{Method \code{saved_search()}}{
Get a specific saved search in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$saved_search(search)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{search}}{A specific saved search in the library.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Only get the saved searches, not search results.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-tags"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-tags}{}}}
\subsection{Method \code{tags()}}{
All tags in the library
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Zotero-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Zotero-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Zotero$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
