% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{zotero_params}
\alias{zotero_params}
\title{Parameters for Zotero API}
\usage{
zotero_params(
  pagination = NULL,
  filter = NULL,
  item_search = NULL,
  tag_search = NULL,
  format = NULL
)
}
\arguments{
\item{pagination}{A \code{\link[=param_pagination]{param_pagination()}} object that defines pagination
settings such as the starting index (\code{start}), the number of results to
return (\code{limit}), and sorting options.}

\item{filter}{A \code{\link[=param_filter]{param_filter()}} object that provides filtering options for
the search results, such as filtering by item keys, item types, or
modification dates.}

\item{item_search}{A \code{\link[=param_search]{param_search()}} object used to refine item searches
based on keywords or other criteria.}

\item{tag_search}{A \code{\link[=param_search]{param_search()}} object used to refine tag searches
based on specific tags or keywords.}

\item{format}{A \code{\link[=param_format]{param_format()}} object that specifies the desired response
format, including options for citation styles, response content, and more.}
}
\description{
This function constructs a set of parameters to be used when making requests
to the Zotero API. It allows users to specify pagination, filtering, search
parameters, and format options.
}
