% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{zotero_perform}
\alias{zotero_perform}
\title{Perform the Zotero API Request}
\usage{
zotero_perform(req, ..., key = NULL)
}
\arguments{
\item{req}{A \code{httr2} request or a list of \code{httr2} request objects created by
\code{\link[=zotero_request]{zotero_request()}}.}

\item{...}{Additional arguments passed to method dispatch.
\itemize{
\item When \code{req} is a single request, these arguments are passed to
\code{\link[httr2:req_perform]{httr2::req_perform()}}.
\item When \code{req} is a list of requests, these arguments are passed to
\code{\link[httr2:req_perform_parallel]{httr2::req_perform_parallel()}}.
}}

\item{key}{Optional authenticated key for the request, if required. By
default, the function will attempt to use the \link[=zotero_key]{global key}.}
}
\value{
The response from the Zotero API.
}
\description{
This function executes a Zotero API request created by \code{zotero_request}. It
uses method dispatch to handle different types of request execution based on
the input. If a single request is provided, it performs the request. If a
list of requests is provided, it performs the requests in parallel.
}
