% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{zotero_revoke}
\alias{zotero_revoke}
\title{Revoke Zotero OAuth Authorization}
\usage{
zotero_revoke(oauth_userid = NULL)
}
\arguments{
\item{oauth_userid}{Optional user ID. If provided, the function will attempt
to find the cached OAuth token for this user. If not provided, the function
will attempt to use the current global OAuth key. For your user ID, visit
\url{https://www.zotero.org/settings/keys}.}
}
\description{
This function revokes the OAuth authorization for a given user. It also
removes the cached OAuth token and deletes any associated token files.
}
